+++ NEWS.md
@@ -10,8 +10,8 @@
 
 * Better handling of missing values to match `R` behavior. In the following
   functions, if there is at least one missing value and `na.rm = FALSE` (the 
-  default), then the output will be `NA`: `"max()"`, `"mean()"`, `"median()"`,
-  `"min()"`, `"sd()"`, `"sum()"`, `"var()"` (#120).
+  default), then the output will be `NA`: `max()`, `mean()`, `median()`, `min()`,
+  `sd()`, `sum()`, `var()` (#120).
   
 * New argument `cluster_with_columns` in `collect()`, `compute()`, and `fetch()`.